Common SearchQuery, SearchQueryLabel, and SearchQueryEqual
