+++ components/common/SearchBar/searchBar.tsx
@@ -4,6 +4,9 @@ import { Autocomplete, InputBase, InputAdornment } from '@mui/material';
 import Popper from '@mui/material/Popper';
 import { Box, Paper } from '@mui/material';
 import { useEffect } from 'react';
+import SearchQuery from '../../../modules/SearchQuery/SearchQuery';
+import searchQueryLabel from '../../../modules/searchQueryLabel/searchQueryLabel';
+import searchQueryEqual from '../../../modules/searchQueryEqual/searchQueryEqual';
 // import { searchAutocomplete } from '../../autocomplete';
 
 /**
@@ -15,13 +18,6 @@ type SearchProps = {
   value: SearchQuery[];
   setValue: Function;
   disabled?: boolean;
-};
-
-type SearchQuery = {
-  prefix?: string;
-  number?: string;
-  professorName?: string;
-  sectionNumber?: string;
 };
 
 /**
@@ -141,21 +137,9 @@ export const SearchBar = (props: SearchProps) => {
               </Box>
             </li>
           )}
-          isOptionEqualToValue={(option, value) => {
-            if (option.prefix !== value.prefix) {
-              return false;
-            }
-            if (option.professorName !== value.professorName) {
-              return false;
-            }
-            if (option.number !== value.number) {
-              return false;
-            }
-            if (option.sectionNumber !== value.sectionNumber) {
-              return false;
-            }
-            return true;
-          }}
+          isOptionEqualToValue={(option, value) =>
+            searchQueryEqual(option, value)
+          }
           PopperComponent={(props) => {
             return (
               <Popper {...props} className="rounded-none" placement="bottom" />
@@ -179,20 +163,3 @@ export const SearchBar = (props: SearchProps) => {
 SearchBar.defaultProps = {
   disabled: true,
 };
-
-function searchQueryLabel(query: SearchQuery): string {
-  let result = '';
-  if (query.prefix !== undefined) {
-    result += query.prefix;
-  }
-  if (query.number !== undefined) {
-    result += ' ' + query.number;
-  }
-  if (query.sectionNumber !== undefined) {
-    result += '.' + query.sectionNumber;
-  }
-  if (query.professorName !== undefined) {
-    result += ' ' + query.professorName;
-  }
-  return result.trim();
-}
