+++ modules/GraphProps/GraphProps.tsx
