+++ modules/SearchQuery/SearchQuery.tsx
@@ -1,0 +1,7 @@
+type SearchQuery = {
+  prefix?: string;
+  number?: string;
+  professorName?: string;
+  sectionNumber?: string;
+};
+export default SearchQuery;
